View refactor
